+++ src/InvoiceData/Service/InvoiceDataServiceInterface.php
@@ -1,0 +1,17 @@
+<?php
+
+/**
+ * Copyright © MB Arbatos Klubas. All rights reserved.
+ * See LICENSE file for license details.
+ */
+
+declare(strict_types=1);
+
+namespace FreshAdvance\Invoice\InvoiceData\Service;
+
+use FreshAdvance\Invoice\InvoiceData\DataType\InvoiceDataInterface;
+
+interface InvoiceDataServiceInterface
+{
+    public function getInvoiceDataByOrderId(string $orderId): InvoiceDataInterface;
+}