+++ tests/Integration/ServiceAvailabilityTest.php
@@ -111,7 +111,7 @@ class ServiceAvailabilityTest extends IntegrationTestCase
             [\FreshAdvance\Invoice\Repository\ShopRepositoryInterface::class],
 
             // todo: move shared service items to where they belong.
-            [\FreshAdvance\Invoice\InvoiceData\Service\Invoice::class],
+            [\FreshAdvance\Invoice\InvoiceData\Service\InvoiceDataServiceInterface::class],
             [\FreshAdvance\Invoice\InvoiceData\Service\InvoiceFileServiceInterface::class],
 
             // Settings